+++ Assets/Scripts/Controllers/BattleScene/BattleMenuDebugController.cs
@@ -1,29 +1,29 @@
+using System.Security.Cryptography;
 using UnityEngine;
 
 namespace Controllers.BattleScene
 {
     public class BattleMenuDebugController : MonoBehaviour
     {
-        // [SerializeField] private GameObject debugButtons;
-        [SerializeField] private BattleMenuController battleMenuController;
-        // private bool debugMode;
+        private BattleMenuController battleMenuController;
+        private StateController stateController;
 
-        // public void ToggleDebugMode()
-        // {
-        //     debugMode = !debugMode;
-        //     debugButtons.SetActive(debugMode);
-        // }
+        private void Start()
+        {
+            battleMenuController = FindObjectOfType<BattleMenuController>();
+            stateController = FindObjectOfType<StateController>();
+        }
 
         public void LoseBattle()
         {
             battleMenuController.soPlayerCharacters.ForEach(c => c.health = 0);
-            battleMenuController.gm.stateController.fsm.ChangeState(StateController.States.ReadyForNextTurn);
+            stateController.fsm.ChangeState(StateController.States.ReadyForNextTurn);
         }
 
         public void WinBattle()
         {
             battleMenuController.soEnemyCharacters.ForEach(c => c.health = 0);
-            battleMenuController.gm.stateController.fsm.ChangeState(StateController.States.ReadyForNextTurn);
+            stateController.fsm.ChangeState(StateController.States.ReadyForNextTurn);
         }
     }
 }
